refactor(frontend): split book-library view into sub-components
What changed: src/views/book-library.tsx (713 lines) split into a thin orchestrator (160 lines) plus src/components/library/library-chrome.tsx (heading + workspace path + totals StatTile grid + filter pills) and src/components/library/library-grid.tsx (skeleton / empty-state / authors→series→BookCard render, including the BookCard internals — cover art, action menu, edit/cover/re-parse/delete modals, paused-snapshot badge — plus the NewBookCard tile). The orchestrator now owns filter state, the workspace-info fetch effect, the allBooks/totals memo, filter-pill labels, and applies the active filter at the boundary so LibraryGrid stays a pure render of pre-filtered authors. Public BookLibraryView prop contract is unchanged — callers in App.tsx / routes need no updates.

Why: unblocks parallel work on the library search/tags filter chrome and the card↔table view toggle without two agents fighting over the render tree of a 713-line view. Wave-0 refactor for the post-v1.3.1 library bundle; mirrors the listen-view split pattern landed in plan 60 (src/views/listen.tsx as a thin orchestrator over three region sub-components under src/components/listen/).

Behaviour-neutral confirmation: no new features, no new state, no new filters, no visual changes. Every data-testid, className, and child order matches the pre-refactor JSX. The existing src/views/book-library.test.tsx (17 cases — loading affordance, populated grid, cover-image overlay, metadata strip, coverFraming, paused/halted badges, hydrate swap) stays green with zero assertion edits. One semantic preservation worth noting: the EmptyLibrary copy still fires only when the underlying library is genuinely empty (authors.length === 0), not when an active filter has hidden every card — passed through LibraryGrid as a dedicated isLibraryEmpty flag distinct from the filtered authors prop.

Verification: npm run verify (cache-aware) green — lint, typecheck, frontend (1140 tests), server (1095 tests + 8 skipped), scripts, sidecar, e2e (44 passed, 2 flaky-then-passed on retry, 2 skipped), build.
